+++ src/Rulyotano.Math.Interpolation.Bezier/BezierCurve.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
@@ -6,19 +6,26 @@ using System.Text;
 
 namespace Rulyotano.Math.Interpolation.Bezier
 {
-    public static class InterpolationExtensionsBezier
+    public class BezierCurve
     {
-        public static string BezierToPath(this IEnumerable<BezierCurveSegment> bezierPaths)
+        public BezierCurve(IEnumerable<BezierCurveSegment> segments)
+        {
+            Segments = segments;
+        }
+
+        public IEnumerable<BezierCurveSegment> Segments { get; }
+
+        public string ToPath()
         {
             var builder = new StringBuilder();
             Func<double, string> round = n => System.Math.Round(n, 3).ToString(CultureInfo.InvariantCulture);
 
-            if (bezierPaths.Any())
+            if (Segments.Any())
             {
-                var head = bezierPaths.First();
+                var head = Segments.First();
                 builder.Append($"M{round(head.StartPoint.X)},{round(head.StartPoint.Y)}");
 
-                foreach (var item in bezierPaths)
+                foreach (var item in Segments)
                 {
                     builder.Append($" C{round(item.FirstControlPoint.X)},{round(item.FirstControlPoint.Y)} {round(item.SecondControlPoint.X)},{round(item.SecondControlPoint.Y)} {round(item.EndPoint.X)},{round(item.EndPoint.Y)}");
                 }