+++ src/Rulyotano.Math.Interpolation.Bezier/Interpolation.Bezier.cs
@@ -9,32 +9,33 @@ namespace Rulyotano.Math.Interpolation.Bezier
         /// <summary>
         /// Interpolate a list of points to a list of ordered Bezier curve segments
         /// </summary>
-        /// <param name="points"></param>
+        /// <param name="pointList"></param>
         /// <param name="isClosedCurve">True if is a closed curve</param>
         /// <param name="smoothValue">Optional value for making the curve smoother or not</param>
         /// <returns></returns>
-        public static IList<BezierCurveSegment> PointsToBezierCurves(IList<Point> points, bool isClosedCurve, double smoothValue = 0.8)
+        public static BezierCurve PointsToBezierCurves(IEnumerable<Point> points, bool isClosedCurve, double smoothValue = 0.8)
         {
-            if (points.Count < 3)
-                return new List<BezierCurveSegment>();
+            var pointList = points.ToList();
+            if (pointList.Count < 3)
+                return new BezierCurve(new List<BezierCurveSegment>());
 
             var toRet = new List<BezierCurveSegment>();
 
             //if is close curve then add the first point at the end
             if (isClosedCurve)
-                points.Add(points.First());
+                pointList.Add(pointList.First());
 
-            for (var i = 0; i < points.Count - 1; i++)   //iterate for points but the last one
+            for (var i = 0; i < pointList.Count - 1; i++)   //iterate for points but the last one
             {
                 // Assume we need to calculate the control
                 // points between (x1,y1) and (x2,y2).
                 // Then x0,y0 - the previous vertex,
                 //      x3,y3 - the next one.
-                var x1 = points[i].X;
-                var y1 = points[i].Y;
+                var x1 = pointList[i].X;
+                var y1 = pointList[i].Y;
 
-                var x2 = points[i + 1].X;
-                var y2 = points[i + 1].Y;
+                var x2 = pointList[i + 1].X;
+                var y2 = pointList[i + 1].Y;
 
                 double x0;
                 double y0;
@@ -43,44 +44,44 @@ namespace Rulyotano.Math.Interpolation.Bezier
                 {
                     if (isClosedCurve)
                     {
-                        var previousPoint = points[points.Count - 2];    //last Point, but one (due inserted the first at the end)
+                        var previousPoint = pointList[pointList.Count - 2];    //last Point, but one (due inserted the first at the end)
                         x0 = previousPoint.X;
                         y0 = previousPoint.Y;
                     }
                     else    //Get some previous point
                     {
-                        var previousPoint = points[i];  //if is the first point the previous one will be it self
+                        var previousPoint = pointList[i];  //if is the first point the previous one will be it self
                         x0 = previousPoint.X;
                         y0 = previousPoint.Y;
                     }
                 }
                 else
                 {
-                    x0 = points[i - 1].X;   //Previous Point
-                    y0 = points[i - 1].Y;
+                    x0 = pointList[i - 1].X;   //Previous Point
+                    y0 = pointList[i - 1].Y;
                 }
 
                 double x3, y3;
 
-                if (i == points.Count - 2)    //if is the last point
+                if (i == pointList.Count - 2)    //if is the last point
                 {
                     if (isClosedCurve)
                     {
-                        var nextPoint = points[1];  //second Point(due inserted the first at the end)
+                        var nextPoint = pointList[1];  //second Point(due inserted the first at the end)
                         x3 = nextPoint.X;
                         y3 = nextPoint.Y;
                     }
                     else    //Get some next point
                     {
-                        var nextPoint = points[i + 1];  //if is the last point the next point will be the last one
+                        var nextPoint = pointList[i + 1];  //if is the last point the next point will be the last one
                         x3 = nextPoint.X;
                         y3 = nextPoint.Y;
                     }
                 }
                 else
                 {
-                    x3 = points[i + 2].X;   //Next Point
-                    y3 = points[i + 2].Y;
+                    x3 = pointList[i + 2].X;   //Next Point
+                    y3 = pointList[i + 2].Y;
                 }
 
                 var xc1 = Numeric.Middle(x0, x1);
@@ -113,13 +114,13 @@ namespace Rulyotano.Math.Interpolation.Bezier
                 toRet.Add(new BezierCurveSegment
                 (
                     startPoint: new Point(x1, y1),
-                    endPoint: new Point(x2, y2),
                     firstControlPoint:i == 0 && !isClosedCurve ? new Point(x1, y1) : new Point(ctrl1X, ctrl1Y),
-                    secondControlPoint: i == points.Count - 2 && !isClosedCurve ? new Point(x2, y2) : new Point(ctrl2X, ctrl2Y)
+                    secondControlPoint: i == pointList.Count - 2 && !isClosedCurve ? new Point(x2, y2) : new Point(ctrl2X, ctrl2Y),
+                    endPoint: new Point(x2, y2)
                 ));
             }
 
-            return toRet;
+            return new BezierCurve(toRet);
         }
     }
 }