+++ src/Rulyotano.Math.Interpolation.Bezier/CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## [Unreleased]
 
-## [1.0.0] - 2022-09-12
-### Added
-- Bezier interpolation method
+## [3.0.0] - 2023-06-24
+### Updated
+- Refactor the result by returning a new BezierCurve class. Also improved some code smells.
 
 ## [2.0.0] - 2022-09-12
 ### Added
 - Extension to convert Bezier collection into SVG compatible path
+
+## [1.0.0] - 2022-09-12
+### Added
+- Bezier interpolation method